Add extensions to aaq, flow_results.
