+++ rdw_ingestion_tools/api/aaq/requests/inbounds.py
@@ -5,7 +5,7 @@ from attrs import define
 from httpx import Client
 from pandas import DataFrame
 
-from .. import get_paginated
+from ..extensions.httpx import get_paginated
 
 
 class FAQModel(TypedDict):
